fix: add to compose details in compose_action
