+++ src/modules/contact.mjs
@@ -1,4 +1,3 @@
-import { mapIter } from "./iter.mjs";
 import {
   categoryStringToArr,
   reduceCategories,
@@ -95,24 +94,4 @@ export function toRFC5322EmailAddress(value) {
     ({ email, name } = value);
   }
   return name ? `${name} <${email}>` : email;
-}
-
-export async function addContactsToComposeDetails(fieldName, tab, contacts) {
-  const details = await browser.compose.getComposeDetails(tab.id);
-  const addresses = details[fieldName];
-  let map = new Map();
-  addresses.forEach((addr) => {
-    const { address, name } = emailAddresses.parseOneAddress(addr);
-    map.set(address, name);
-  });
-  contacts.forEach(({ email, name }) => {
-    // Add this contact if it doesn't exist in the map
-    if (email != null && !map.has(email)) map.set(email, name);
-  });
-  const emailList = [...mapIter(map.entries(), toRFC5322EmailAddress)];
-  // set compose details
-  await browser.compose.setComposeDetails(tab.id, {
-    ...details,
-    [fieldName]: emailList,
-  });
 }
